Define a plugin for object migration.
