+++ index.html
@@ -81,6 +81,7 @@
         openmct.install(openmct.plugins.Tabs());
         openmct.install(openmct.plugins.FlexibleLayout());
         openmct.install(openmct.plugins.LADTable());
+        openmct.install(openmct.plugins.ObjectMigration());
         openmct.start();
     </script>
 </html>
